# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Learning Management System (LMS) API
  description: |
    API for managing a Learning Management System with courses, modules, lessons, and user tracking.
    This API provides comprehensive functionality for creating and managing educational content,
    user enrollments, and progress tracking.
  version: 1.0.0
  contact:
    name: LMS API Support
    email: support@lmsapi.example.com
servers:
  - url: https://api.lms.example.com/v1
    description: Production server
  - url: https://staging-api.lms.example.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
    Error:
      type: object
      properties:
        id:
          type: string
          example: api.course.get
        ver:
          type: string
          example: "1.0"
        ts:
          type: string
          format: date-time
        params:
          type: object
          properties:
            resmsgid:
              type: string
              format: uuid
            status:
              type: string
              enum: [failed]
            err:
              type: string
              example: RESOURCE_NOT_FOUND
            errmsg:
              type: string
              example: Course not found
        responseCode:
          type: integer
          example: 404
        result:
          type: object
...
    Lesson:
      type: object
      properties:
        lessonId:
          type: string
          format: uuid
        tenentId:
          type: string
          format: uuid
        title:
          type: string
        alias:
          type: string
        description:
          type: string
        image:
          type: string
          nullable: true
        startDatetime:
          type: string
          format: date-time
        endDatetime:
          type: string
          format: date-time
        storage:
          type: string
          enum: [local, cloud]
        noOfAttempts:
          type: string
        attemptsGrade:
          type: string
          enum: [highest, average, latest]
        format:
          type: string
          enum: [video, document, quiz, scorm, html]
        mediaId:
          type: string
          format: uuid
          nullable: true
        eligibilityCriteria:
          type: string
          nullable: true
        idealTime:
          type: integer
        resume:
          type: boolean
        totalMarks:
          type: integer
        passingMarks:
          type: integer
        status:
          type: string
          enum: [draft, published, unpublished]
        params:
          type: object
          additionalProperties: true
        mediaContent:
          type: object
          properties:
            format:
              type: string
            subFormat:
              type: string
            source:
              type: string
            params:
              type: object
              additionalProperties: true
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - title
        - description
...
    SuccessResponse:
      type: object
      properties:
        id:
          type: string
        ver:
          type: string
        ts:
          type: string
          format: date-time
        params:
          type: object
          properties:
            resmsgid:
              type: string
              format: uuid
            status:
              type: string
              enum: [successful]
            err:
              type: object
            errmsg:
              type: string
        responseCode:
          type: integer
        result:
          type: object
...
security:
  - bearerAuth: []
paths:
...
  /lessons:
    post:
      summary: Create a new lesson
      description: Creates a new lesson
      operationId: createLesson
      tags:
        - Lessons
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                alias:
                  type: string
                description:
                  type: string
                image:
                  type: string
                  format: binary
                startDatetime:
                  type: string
                  format: date-time
                endDatetime:
                  type: string
                  format: date-time
                storage:
                  type: string
                  enum: [local, cloud]
                noOfAttempts:
                  type: string
                attemptsGrade:
                  type: string
                  enum: [highest, average, latest]
                format:
                  type: string
                  enum: [video, document, quiz, scorm, html]
                eligibilityCriteria:
                  type: string
                idealTime:
                  type: integer
                resume:
                  type: boolean
                totalMarks:
                  type: integer
                passingMarks:
                  type: integer
                status:
                  type: string
                  enum: [draft, published, unpublished]
                mediaContent:
                  type: string
                  description: JSON string of media content details
              required:
                - title
                - description
      responses:
        '201':
          description: Lesson created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      responseCode:
                        type: integer
                        example: 201
                      result:
                        $ref: '#/components/schemas/Lesson'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    get:
      summary: Get all lessons
      description: Retrieves all lessons
      operationId: getLessons
      tags:
        - Lessons
      parameters:
        - name: status
          in: query
          description: Filter by lesson status
          schema:
            type: string
            enum: [draft, published, unpublished]
        - name: format
          in: query
          description: Filter by lesson format
          schema:
            type: string
            enum: [video, document, quiz, scorm, html]
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      result:
                        type: array
                        items:
                          $ref: '#/components/schemas/Lesson'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /lessons/{lessonId}:
    get:
      summary: Get lesson by ID
      description: Retrieves a specific lesson by ID
      operationId: getLessonById
      tags:
        - Lessons
      parameters:
        - name: lessonId
          in: path
          required: true
          description: UUID of the lesson
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      result:
                        $ref: '#/components/schemas/Lesson'
        '404':
          description: Lesson not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      summary: Update lesson
      description: Updates an existing lesson
      operationId: updateLesson
      tags:
        - Lessons
      parameters:
        - name: lessonId
          in: path
          required: true
          description: UUID of the lesson
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                alias:
                  type: string
                description:
                  type: string
                image:
                  type: string
                  format: binary
                startDatetime:
                  type: string
                  format: date-time
                endDatetime:
                  type: string
                  format: date-time
                storage:
                  type: string
                  enum: [local, cloud]
                noOfAttempts:
                  type: string
                attemptsGrade:
                  type: string
                  enum: [highest, average, latest]
                format:
                  type: string
                  enum: [video, document, quiz, scorm, html]
                eligibilityCriteria:
                  type: string
                idealTime:
                  type: integer
                resume:
                  type: boolean
                totalMarks:
                  type: integer
                passingMarks:
                  type: integer
                status:
                  type: string
                  enum: [draft, published, unpublished]
                mediaContent:
                  type: string
                  description: JSON string of media content details
      responses:
        '200':
          description: Lesson updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      result:
                        $ref: '#/components/schemas/Lesson'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Lesson not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      summary: Delete lesson
      description: Deletes a lesson
      operationId: deleteLesson
      tags:
        - Lessons
      parameters:
        - name: lessonId
          in: path
          required: true
          description: UUID of the lesson
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lesson deleted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      result:
                        type: object
                        properties:
                          message:
                            type: string
                            example: Lesson deleted successfully
        '404':
          description: Lesson not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /courses/{courseId}/modules/{moduleId}/lessons/{lessonId}:
    delete:
      summary: Remove lesson from course/module
      description: Removes a lesson from a course/module
      operationId: removeLessonFromCourseModule
      tags:
        - Lessons
      parameters:
        - name: courseId
          in: path
          required: true
          description: UUID of the course
          schema:
            type: string
            format: uuid
        - name: moduleId
          in: path
          required: true
          description: UUID of the module (use "null" for course-level lessons)
          schema:
            type: string
        - name: lessonId
          in: path
          required: true
          description: UUID of the lesson
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lesson removed from course/module successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      result:
                        type: object
                        properties:
                          message:
                            type: string
                            example: Lesson removed from course/module successfully
        '404':
          description: Course, module, or lesson not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...